base chart

+++ app/src/main/java/com/example/testcompose/ui/DataModel.kt
@@ -1,4 +1,4 @@
 package com.example.testcompose.ui
 
-data class DataModel(val year: Int, val index: Int) {
+data class DataModel(val year: Int, val index: Float) {
 }
